+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added default config for copyPublicDir and manifest
+
 ## v0.1.1 (2025-03-05)
 * Improved package file